Split samplers implementation across different files
