+++ src/torchcodec/samplers/__init__.py
@@ -1,6 +1,2 @@
-from ._implementation import (
-    clips_at_random_indices,
-    clips_at_random_timestamps,
-    clips_at_regular_indices,
-    clips_at_regular_timestamps,
-)
+from ._index_based import clips_at_random_indices, clips_at_regular_indices
+from ._time_based import clips_at_random_timestamps, clips_at_regular_timestamps
